Add filename detectors and activationEvents
JenkinsFile and JobDSL filename detectors since that's what I wanted it 
for in the first place. Doh!

For activationEvents for the LSP though I'm not considering Jenkinsfile 
or JobDSL as  those will just report a bunch of errors. Hence why also 
not doing an `onLanguage:Groovy`.

I guess I need a way to disable the LSP.